refactor: AuthService에서 MemberService 의존성 분리 설계
