+++ src/main/java/com/depromeet/stonebed/domain/auth/application/AuthService.java
@@ -8,10 +8,12 @@ import com.depromeet.stonebed.domain.auth.dto.request.RefreshTokenRequest;
 import com.depromeet.stonebed.domain.auth.dto.response.AuthTokenResponse;
 import com.depromeet.stonebed.domain.auth.dto.response.SocialClientResponse;
 import com.depromeet.stonebed.domain.auth.dto.response.TokenPairResponse;
-import com.depromeet.stonebed.domain.member.application.MemberService;
+import com.depromeet.stonebed.domain.member.dao.MemberRepository;
 import com.depromeet.stonebed.domain.member.domain.Member;
 import com.depromeet.stonebed.domain.member.domain.MemberRole;
+import com.depromeet.stonebed.domain.member.domain.Profile;
 import com.depromeet.stonebed.domain.member.dto.request.CreateMemberRequest;
+import com.depromeet.stonebed.domain.member.dto.request.NicknameCheckRequest;
 import com.depromeet.stonebed.global.error.ErrorCode;
 import com.depromeet.stonebed.global.error.exception.CustomException;
 import com.depromeet.stonebed.global.util.MemberUtil;
@@ -29,9 +31,9 @@ public class AuthService {
 
     private final AppleClient appleClient;
     private final KakaoClient kakaoClient;
-    private final MemberService memberService;
     private final JwtTokenService jwtTokenService;
     private final MemberUtil memberUtil;
+    private final MemberRepository memberRepository;
 
     public SocialClientResponse authenticateFromProvider(OAuthProvider provider, String token) {
         /* token
@@ -46,7 +48,9 @@ public class AuthService {
 
     public AuthTokenResponse socialLogin(
             OAuthProvider oAuthProvider, String oauthId, String email) {
-        Optional<Member> memberOptional = memberService.getMemberByOauthId(oAuthProvider, oauthId);
+        Optional<Member> memberOptional =
+                memberRepository.findByOauthInfoOauthProviderAndOauthInfoOauthId(
+                        oAuthProvider.getValue(), oauthId);
 
         return memberOptional
                 .map(
@@ -64,7 +68,9 @@ public class AuthService {
                         () -> {
                             // 회원가입이 안된 경우, 임시 회원가입 진행
                             Member newMember =
-                                    memberService.socialSignUp(oAuthProvider, oauthId, email);
+                                    Member.createOAuthMember(oAuthProvider, oauthId, email);
+                            memberRepository.save(newMember);
+
                             // 임시 토큰 발행
                             TokenPairResponse temporaryTokenPair =
                                     jwtTokenService.generateTemporaryTokenPair(newMember);
@@ -78,9 +84,14 @@ public class AuthService {
         Member currentMember = memberUtil.getCurrentMember();
         // 사용자 회원가입
         if (memberUtil.getMemberRole().equals(MemberRole.TEMPORARY.getValue())) {
-            Member member = memberService.registerMember(currentMember, request);
+            // 닉네임 검증
+            memberUtil.checkNickname(NicknameCheckRequest.of(request.nickname()));
+
+            // 명시적 변경 감지
+            Member registerMember = registerMember(currentMember, request);
+
             // 새 토큰 생성
-            TokenPairResponse tokenPair = getLoginResponse(member);
+            TokenPairResponse tokenPair = getLoginResponse(registerMember);
             return AuthTokenResponse.of(tokenPair, false);
         }
         throw new CustomException(ErrorCode.ALREADY_EXISTS_MEMBER);
@@ -108,9 +119,7 @@ public class AuthService {
         return jwtTokenService.generateTokenPair(member.getId(), MemberRole.TEMPORARY);
     }
 
-    public void withdraw(
-            /** OAuthProvider provider */
-            ) {
+    public void withdraw() {
         Member member = memberUtil.getCurrentMember();
         /**
          * TODO: 런칭데이 이후 고도화 if (provider.equals(OAuthProvider.APPLE)) {
@@ -118,5 +127,15 @@ public class AuthService {
          */
         jwtTokenService.deleteRefreshToken(member.getId());
         member.withdrawal();
+    }
+
+    private Member registerMember(Member member, CreateMemberRequest request) {
+        member.updateProfile(
+                Profile.createProfile(
+                        request.nickname(), member.getProfile().getProfileImageUrl()));
+        member.updateRaisePet(request.raisePet());
+        member.updateMemberRole(MemberRole.USER);
+        memberRepository.save(member);
+        return member;
     }
 }
