+++ src/main/java/com/depromeet/stonebed/global/util/MemberUtil.java
@@ -2,6 +2,7 @@ package com.depromeet.stonebed.global.util;
 
 import com.depromeet.stonebed.domain.member.dao.MemberRepository;
 import com.depromeet.stonebed.domain.member.domain.Member;
+import com.depromeet.stonebed.domain.member.dto.request.NicknameCheckRequest;
 import com.depromeet.stonebed.global.error.ErrorCode;
 import com.depromeet.stonebed.global.error.exception.CustomException;
 import lombok.RequiredArgsConstructor;
@@ -30,7 +31,20 @@ public class MemberUtil {
         return securityUtil.getCurrentMemberRole();
     }
 
-    public String getMemberProvider() {
-        return getCurrentMember().getOauthInfo().getOauthProvider();
+    public void checkNickname(NicknameCheckRequest request) {
+        validateNicknameNotDuplicate(request.nickname());
+        if (validateNicknameText(request.nickname())) {
+            throw new CustomException(ErrorCode.MEMBER_INVALID_NICKNAME);
+        }
+    }
+
+    private boolean validateNicknameText(String nickname) {
+        return nickname == null || nickname.length() < 2 || nickname.length() > 14;
+    }
+
+    private void validateNicknameNotDuplicate(String nickname) {
+        if (memberRepository.existsByProfileNickname(nickname)) {
+            throw new CustomException(ErrorCode.MEMBER_ALREADY_NICKNAME);
+        }
     }
 }
